requirements.txt
numpy
pandas
streamlit
python-dotenv
streamlit_option_menu
streamlit_extras
PyPDF2
langchain
openai
tiktoken
faiss-cpu
